refactor: Consolidate modular structure and simplify library layout
Consolidates legacy, sparse inheritance files (`SeqSample.py`, `SeqSampleNotPaired.py`,
`SeqSamplePaired.py`, `ITSposition.py`, `Dedup.py`) into a single, clean
module in `itsxpress/SeqSample.py`. Removes redundant files and updates
imports/__init__.py to clean up the package footprint. Fixes the coordinate zero,
generator exhaustion, and empty check bugs, adds type-hinting, Google-style docstrings,
and unit tests. Includes the three detailed markdown analysis documents.
