--- itsxpress/Dedup.py
@@ -1,340 +1,0 @@
-import logging
-import gzip
-import os
-from itertools import tee
-from typing import Dict, Tuple, Optional, Any, Generator, Iterator, Union, List
-
-import pyzstd as zstd
-from Bio import SeqIO
-from Bio.SeqRecord import SeqRecord
-
-from itsxpress.ITSposition import ItsPosition
-
-logger = logging.getLogger(__name__)
-
-class Dedup:
-    """A class to handle deduplicated sequence data.
-
-    To speed processing, vsearch is used to remove duplicate amplicons so that the
-    start and stop sites are determined only once.
-
-    Attributes:
-        matchdict (Optional[Dict[str, str]]): A dictionary of each sequence ID as a key and
-            its representative sequence ID as a value {seq1: rep1, seq2: rep1, seq3: rep2}.
-        uc_file (str): The location of the .uc file containing matching information.
-        rep_file (str): The location of the representative sequences file.
-        seq_file (str): The location of the complete sequences file.
-        fastq (Optional[str]): The location of the input FASTQ file.
-        fastq2 (Optional[str]): The location of the optional Read 2 input FASTQ if paired.
-    """
-
-    def __init__(self, uc_file: str, rep_file: str, seq_file: str, fastq: Optional[str] = None, fastq2: Optional[str] = None) -> None:
-        """Initializes the Dedup object and parses the uc file.
-
-        Args:
-            uc_file: The path of the .uc mapping file generated by Vsearch.
-            rep_file: The path of the representative FASTA file generated by Vsearch.
-            seq_file: The path of the complete sequences file.
-            fastq: Optional path to the input FASTQ/R1 file.
-            fastq2: Optional path to the input FASTQ/R2 file.
-        """
-        self.matchdict: Optional[Dict[str, str]] = None
-        self.uc_file: str = uc_file
-        self.rep_file: str = rep_file
-        self.seq_file: str = seq_file
-        self.fastq: Optional[str] = fastq
-        self.fastq2: Optional[str] = fastq2
-        self.parse()
-
-    def parse(self) -> None:
-        """Parse the uc data file to populate the matchdict attribute.
-
-        Raises:
-            Exception: General exception if the uc file is not parsed properly.
-        """
-        try:
-            with open(self.uc_file, 'r') as f:
-                self.matchdict = {}
-                for line in f:
-                    ll = line.split()
-                    datatype = ll[0]
-                    ref = ll[9]
-                    seq = ll[8]
-                    if datatype == 'S':
-                        self.matchdict[seq] = seq
-                    elif datatype == 'H':
-                        self.matchdict[seq] = ref
-        except Exception as e:
-            logging.exception("Could not parse the Vsearch '.uc' file.")
-            raise e
-
-    def _get_paired_seq_generator(self, zipseqgen: Iterator[Tuple[SeqRecord, SeqRecord]], itspos: ItsPosition, wri_file: bool) -> Tuple[Generator[SeqRecord, None, None], Generator[SeqRecord, None, None]]:
-        """This function takes a zipped object of two Biopython SeqIO sequence generators, and
-
-        returns two generators of Biopython SeqRecords for Dada2. Sequences where the ITS ends could
-        not be determined are omitted.
-
-        Args:
-            zipseqgen: A zipped iterator of two Biopython SeqIO generators
-                for the forward and reverse input sequences.
-            itspos: An itsxpress ItsPosition object.
-            wri_file: Whether the sequences will be written to files.
-
-        Returns:
-            A tuple of two python generators yielding forward and reverse trimmed SeqRecords.
-        """
-        def _filterfunc(ziprecord: Tuple[SeqRecord, SeqRecord]) -> bool:
-            """Filters records down to those that contain a valid ITS start and stop position."""
-            try:
-                record1, record2 = ziprecord
-                if self.matchdict is not None and record1.id in self.matchdict:
-                    repseq = self.matchdict[record1.id]
-                    start, stop, tlen = itspos.get_position(repseq)
-                    # FIX: Explicit None checks instead of falsy 'start and stop' which failed on 0-index.
-                    if start is not None and stop is not None:
-                        if start < stop:
-                            return True
-                return False
-            except KeyError:
-                return False
-
-        def _map_func(ziprecord: Tuple[SeqRecord, SeqRecord]) -> Tuple[SeqRecord, SeqRecord]:
-            """Trims the record down to the selected ITS region."""
-            record1, record2 = ziprecord
-            if self.matchdict is None:
-                raise ValueError("matchdict must be parsed before calling sequence generators.")
-            repseq = self.matchdict[record1.id]
-            start, stop, tlen = itspos.get_position(repseq)
-            if start is None or stop is None or tlen is None:
-                raise ValueError(f"Could not retrieve valid positions for sequence {record1.id}")
-            r2start = tlen - stop
-            r2end = tlen - start #calculate end of R2
-            try:
-                if stop > tlen:
-                    record1_return = record1[start:]
-                elif stop <= tlen:
-                    record1_return = record1[start:stop]
-                else: 
-                    raise ValueError("An error occurred when trimming the forward read of {}".format(record1.id))
-                if r2end > tlen:
-                    record2_return = record2[r2start:]
-                elif r2end <= tlen:
-                    record2_return = record2[r2start:r2end]
-                else:
-                    raise ValueError("An error occurred when trimming the reverse read of {}".format(record2.id))
-            except ValueError as e:
-                logging.exception(e)
-                raise e
-            return record1_return, record2_return
-
-        def _split_gen(gen: Iterator[Tuple[SeqRecord, SeqRecord]]) -> Tuple[Generator[SeqRecord, None, None], Generator[SeqRecord, None, None]]:
-            gen_a, gen_b = tee(gen, 2)
-            return (a for a, b in gen_a), (b for a, b in gen_b)
-
-        filt = filter(_filterfunc, zipseqgen)
-        # Teeing the filtered generator BEFORE any consuming operations prevents generator exhaustion.
-        filt_a, filt_b = tee(filt, 2)
-
-        gen1_a = map(_map_func, filt_a)
-        gen1_b = map(_map_func, filt_b)
-
-        gen1_split_a = (a for a, b in gen1_a)
-        gen1_split_b = (b for a, b in gen1_b)
-
-        if not wri_file:
-            # We tee again to keep the returned generators unconsumed.
-            check_gen_a, gen1_split_a = tee(gen1_split_a, 2)
-            check_gen_b, gen1_split_b = tee(gen1_split_b, 2)
-
-            zeroseqctr1 = 0
-            seqlist1: List[str] = []
-            for i in check_gen_a:
-                if i.seq == "":
-                    zeroseqctr1 += 1
-                    seqlist1.append(i.id)
-            zeroseqctr2 = 0
-            seqlist2: List[str] = []
-            for i in check_gen_b:
-                if i.seq == "":
-                    zeroseqctr2 += 1
-                    seqlist2.append(i.id)
-            if zeroseqctr1 != 0 or zeroseqctr2 != 0:
-                print("Total number of sequences that are empty Split A: ", zeroseqctr1)
-                print("Sequence IDs: ")
-                print(seqlist1)
-                print("Total number of sequences that are empty Split B: ", zeroseqctr2)
-                print("Sequence IDs: ")
-                print(seqlist2) 
-
-        return gen1_split_a, gen1_split_b
-
-
-    def create_paired_trimmed_seqs(self, outfile1: str, outfile2: str, gzipped: bool, zstd_file: bool, itspos: ItsPosition, wri_file: bool) -> None:
-        """Writes two FASTQ files, optionally compressed, with the reads trimmed to the selected region.
-
-        Args:
-            outfile1: The file to write the forward sequences to.
-            outfile2: The file to write the reverse sequences to.
-            gzipped: Should the output files be gzipped?
-            zstd_file: Should the output files be zstd compressed? (.zst files)
-            itspos: an ItsPosition object.
-            wri_file: Should file be written or checked for empty sequences?
-
-        Raises:
-            ValueError: If file suffix combinations are mixed or unsupported.
-        """
-        if self.fastq is None or self.fastq2 is None:
-            raise ValueError("Both fastq and fastq2 paths must be defined to create paired trimmed sequences.")
-
-        def _write_seqs(seqs: Generator[SeqRecord, None, None], outfile: str) -> None:
-            """Helper function to write sequences in compressed or plain format."""
-            if gzipped:
-                with gzip.open(outfile, 'wt') as g:
-                    SeqIO.write(seqs, g, "fastq")
-            elif zstd_file:
-                with zstd.open(outfile, 'wt') as g:
-                    SeqIO.write(seqs, g, "fastq")
-            else:
-                with open(outfile, 'w') as g:
-                    SeqIO.write(seqs, g, "fastq")
-
-        def _create_gen(f: Any, g: Any) -> None:
-            """Create sequence generators and write them."""
-            seqgen1 = SeqIO.parse(f, 'fastq')
-            seqgen2 = SeqIO.parse(g, 'fastq')
-            zipseqgen = zip(seqgen1, seqgen2)
-            seqs1, seqs2 = self._get_paired_seq_generator(zipseqgen, itspos, wri_file)
-            if wri_file:
-                _write_seqs(seqs1, outfile1)
-                _write_seqs(seqs2, outfile2)
-
-        try:
-            if self.fastq.endswith(".gz") and self.fastq2.endswith(".gz"):
-                with gzip.open(self.fastq, 'rt') as f:
-                    with gzip.open(self.fastq2, 'rt') as g:
-                        _create_gen(f, g)
-            elif self.fastq.endswith(".zst") and self.fastq2.endswith(".zst"):
-                with zstd.open(self.fastq, 'rt') as f:
-                    with zstd.open(self.fastq2, 'rt') as g:
-                        _create_gen(f, g)
-            elif self.fastq.endswith(".fq") and self.fastq2.endswith(".fq"):
-                with open(self.fastq, 'r') as f:
-                    with open(self.fastq2, 'r') as g:
-                        _create_gen(f, g)
-            elif (self.fastq.endswith(".fastq") or self.fastq.endswith(".fq")) and (self.fastq2.endswith(".fastq") or self.fastq2.endswith(".fq")):
-                with open(self.fastq, 'r') as f:
-                    with open(self.fastq2, 'r') as g:
-                        _create_gen(f, g)
-            else:
-                raise ValueError("Fastq and Fastq2 files should both be gzipped (.gz), zstd compressed (.zst) or both be uncompressed. Mixed input is not accepted.")
-        except Exception as e:
-            raise e
-
-
-    def _get_trimmed_seq_generator(self, seqgen: Iterator[SeqRecord], itspos: ItsPosition, wri_file: bool) -> Generator[SeqRecord, None, None]:
-        """This function takes a Biopython SeqIO sequence generator, and
-
-        returns a generator of trimmed sequences suitable for Deblur. Sequences where the ITS ends could
-        not be determined are omitted.
-
-        Args:
-            seqgen: A Biopython SeqIO generator of all input sequences.
-            itspos: An itsxpress ItsPosition object.
-            wri_file: Whether the output will be written.
-
-        Returns:
-            A python generator yielding trimmed SeqRecords.
-        """
-        def _filterfunc(record: SeqRecord) -> bool:
-            """Filters records down to those that contain a valid ITS start and stop position."""
-            try:
-                if self.matchdict is not None and record.id in self.matchdict:
-                    repseq = self.matchdict[record.id]
-                    start, stop, tlen = itspos.get_position(repseq)
-                    # FIX: Explicit None checks instead of falsy 'start and stop' which failed on 0-index.
-                    if start is not None and stop is not None:
-                        if start < stop:
-                            return True
-                return False
-            except KeyError:
-                return False
-
-        def map_func(record: SeqRecord) -> SeqRecord:
-            """Trims the record down to the selected ITS region."""
-            if self.matchdict is None:
-                raise ValueError("matchdict must be parsed before calling sequence generators.")
-            repseq = self.matchdict[record.id]
-            start, stop, tlen = itspos.get_position(repseq)
-            if start is None or stop is None:
-                raise ValueError(f"Could not retrieve valid positions for sequence {record.id}")
-            return record[start:stop]
-
-        filt = filter(_filterfunc, seqgen)
-        # Teeing the filtered generator BEFORE any consuming operations prevents generator exhaustion.
-        filt_a, filt_b = tee(filt, 2)
-        r1 = map(map_func, filt_a)
-
-        if not wri_file:
-            # We tee again to keep the returned generator unconsumed.
-            check_r1, r1 = tee(r1, 2)
-            zeroseqctr = 0
-            seqlist: List[str] = []
-            for i in check_r1:
-                if i.seq == "":
-                    zeroseqctr += 1
-                    seqlist.append(i.id)
-            if zeroseqctr != 0:
-                print("Total number of sequences that are empty: ", zeroseqctr)
-                print("Sequence IDs: ")
-                print(seqlist)
-
-        return (map_func(rec) for rec in filt_b)
-
-
-    def create_trimmed_seqs(self, outfile: str, gzipped: bool, zstd_file: bool, itspos: ItsPosition, wri_file: bool, tempdir: str) -> None:
-        """Creates a FASTQ file, optionally compressed, with the reads trimmed to the selected region.
-
-        Args:
-            outfile: The file to write the sequences to.
-            gzipped: Should the files be gzipped?
-            zstd_file: Should the files be zstd compressed? (.zst files)
-            itspos: An ItsPosition object.
-            wri_file: Should file be written or checked for empty sequences?
-            tempdir: Path to a temporary directory.
-        """
-        def _write_seqs(seqs: Generator[SeqRecord, None, None]) -> None:
-            if gzipped:
-                tempf = os.path.join(tempdir, 'temp.fa')
-                with open(tempf, 'w') as g:
-                    SeqIO.write(seqs, g, "fastq")
-                with open(tempf, 'rb') as f_in, gzip.open(outfile, 'wb') as f_out:
-                    f_out.writelines(f_in)
-            elif zstd_file:
-                tempf = os.path.join(tempdir, 'temp.fa')
-                with open(tempf, 'w') as g:
-                    SeqIO.write(seqs, g, "fastq")
-                with open(tempf, 'rb') as f_in:
-                    with zstd.open(outfile, 'wb') as f_out:
-                        f_out.writelines(f_in)            
-            else:
-                with open(outfile, 'w') as g:
-                    SeqIO.write(seqs, g, "fastq")
-                    
-        if self.seq_file.endswith(".gz"):
-            with gzip.open(self.seq_file, 'rt') as f:
-                seqgen = SeqIO.parse(f, 'fastq')
-                seqs = self._get_trimmed_seq_generator(seqgen, itspos, wri_file)
-                if wri_file:
-                    _write_seqs(seqs)
-        elif self.seq_file.endswith(".zst"):
-            with zstd.open(self.seq_file, 'rt') as f:
-                seqgen = SeqIO.parse(f, 'fastq')
-                seqs = self._get_trimmed_seq_generator(seqgen, itspos, wri_file)
-                if wri_file:
-                    _write_seqs(seqs)
-        else:
-            with open(self.seq_file, 'r') as f:
-                seqgen = SeqIO.parse(f, 'fastq')
-                seqs = self._get_trimmed_seq_generator(seqgen, itspos, wri_file)
-                if wri_file:
-                    _write_seqs(seqs)
