--- itsxpress/ITSposition.py
@@ -1,120 +1,0 @@
-import logging
-from typing import Dict, Tuple, Optional, Any
-
-logger = logging.getLogger(__name__)
-
-class ItsPosition:
-    """Class for ITS positional information derived from hmmsearch domtable files.
-
-    Attributes:
-        domtable (str): Path to the HMMER3 hmmsearch domtable output file.
-        region (str): The region of the ITS to extract. Must be "ITS1", "ITS2", or "ALL".
-        ddict (Dict[str, Any]): A dictionary holding the scores and start/stop positions.
-            Example: {sample: {left: {score: 31, to_pos: 15, from_pos: 0}, right: {score: 32, to_pos: 354, from_pos: 200}, tlen: 449}}
-        leftprefix (str): Prefix of the HMM profiles corresponding to the left boundary.
-        rightprefix (str): Prefix of the HMM profiles corresponding to the right boundary.
-    """
-
-    def __init__(self, domtable: str, region: str) -> None:
-        """Initializes ItsPosition, parses the domtable file, and determines boundaries.
-
-        Args:
-            domtable: Path to the HMMER3 hmmsearch domtable output file.
-            region: The region of the ITS to extract (choices: "ITS1", "ITS2", "ALL").
-        """
-        self.domtable: str = domtable
-        self.ddict: Dict[str, Any] = {}
-        if region == "ITS2":
-            self.leftprefix: str = '3_'
-            self.rightprefix: str = '4_'
-        elif region == "ITS1":
-            self.leftprefix = '1_'
-            self.rightprefix = '2_'
-        elif region == "ALL":
-            self.leftprefix = '1_'
-            self.rightprefix = '4_'
-        self.parse()
-
-    def _score(self, sequence: str, stype: str, score: float, from_pos: int, to_pos: int, tlen: int) -> None:
-        """Evaluates scores and positions from a domtable line and updates ddict.
-
-        Args:
-            sequence: The name/ID of the sequence.
-            stype: Either 'left' or 'right' indicating the boundary type.
-            score: The bit score of the match from hmmsearch.
-            from_pos: The beginning coordinate of the HMM match.
-            to_pos: The ending coordinate of the HMM match.
-            tlen: The total length of the sequence.
-        """
-        if stype in self.ddict[sequence]:
-            if score > self.ddict[sequence][stype]["score"]:
-                self.ddict[sequence][stype]["score"] = score
-                self.ddict[sequence][stype]["to_pos"] = to_pos
-                self.ddict[sequence][stype]["from_pos"] = from_pos
-        else:
-            self.ddict[sequence][stype] = {}
-            self.ddict[sequence][stype]["score"] = score
-            self.ddict[sequence][stype]["to_pos"] = to_pos
-            self.ddict[sequence][stype]["from_pos"] = from_pos
-            self.ddict[sequence]["tlen"] = tlen
-
-    def parse(self) -> None:
-        """Parses the domtable file generated by hmmsearch.
-
-        Identifies matches for left and right HMM boundaries and updates `ddict`
-        with the highest-scoring boundaries.
-
-        Raises:
-            Exception: If an error occurs when parsing HMMER hmmsearch results.
-        """
-        try:
-            with open(self.domtable, 'r') as f:
-                for line in f:
-                    if not line.startswith("#"):
-                        ll = line.split()
-                        sequence = ll[0]
-                        hmmprofile = ll[3]
-                        score = float(ll[13])
-                        from_pos = int(ll[19])
-                        to_pos = int(ll[20])
-                        tlen = int(ll[2])
-                        if sequence not in self.ddict:
-                            self.ddict[sequence] = {}
-                        if hmmprofile.startswith(self.leftprefix):
-                            self._score(sequence, 'left', score, from_pos, to_pos, tlen)
-                        elif hmmprofile.startswith(self.rightprefix):
-                            self._score(sequence, 'right', score, from_pos, to_pos, tlen)
-        except Exception as e:
-            logging.error("Exception occurred when parsing HMMSearch results")
-            raise e
-
-    def get_position(self, sequence: str) -> Tuple[Optional[int], Optional[int], Optional[int]]:
-        """Returns the start and stop positions for a given sequence.
-
-        Args:
-            sequence: The name/ID of the sequence.
-
-        Returns:
-            A tuple of (start coordinate, stop coordinate, total sequence length).
-            Coordinates are 0-indexed and can be None if not found.
-
-        Raises:
-            KeyError: If the input sequence was not parsed or has no identified boundaries.
-        """
-        try:
-            if "left" in self.ddict[sequence]:
-                start = int(self.ddict[sequence]["left"]["to_pos"])
-            else:
-                start = None
-            if "right" in self.ddict[sequence]:
-                stop = int(self.ddict[sequence]["right"]["from_pos"]) - 1
-            else:
-                stop = None
-            if "tlen" in self.ddict[sequence]:
-                tlen = int(self.ddict[sequence]["tlen"])
-            else:
-                tlen = None
-            return (start, stop, tlen)
-        except KeyError:
-            logging.debug("No ITS stop or start sites were identified for sequence {}, skipping.".format(sequence))
-            raise KeyError
