+++ itsxpress/__init__.py
@@ -1,10 +1,6 @@
 import itsxpress.main
 import itsxpress.definitions
-import itsxpress.ITSposition
-import itsxpress.Dedup
 import itsxpress.SeqSample
-import itsxpress.SeqSampleNotPaired
-import itsxpress.SeqSamplePaired
 
 try:
     import itsxpress.q2_itsxpress
